chore: Move all react components into a single package
