+++ package.json
@@ -7,11 +7,8 @@
   "workspaces": ["packages/*"],
   "scripts": {
     "docs": "storybook dev -p 6006",
-    "build": "yarn build:styles && yarn build:icons && yarn build:elements && yarn build:flexbox && yarn build:menu",
+    "build": "yarn build:styles && yarn build:elements",
     "build:styles": "lerna run --scope @seed-ui/styles --stream build",
-    "build:icons": "lerna run --scope @seed-ui/icons --stream build",
-    "build:menu": "lerna run --scope @seed-ui/menu --stream build",
-    "build:flexbox": "lerna run --scope @seed-ui/flexbox --stream build",
     "build:elements": "lerna run --scope @seed-ui/elements --stream build",
     "check:licenses": "license-checker --production --summary --failOn=GPLv3",
     "format": "yarn format:ts && yarn format:js && yarn format:json && yarn format:yaml",
