+++ packages/elements/package.json
@@ -36,10 +36,12 @@
   "author": "Sergey Murawsky",
   "license": "MIT",
   "peerDependencies": {
+    "@seed-ui/styles": "^0.10.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@seed-ui/styles": "^0.10.0",
     "@types/lodash": "^4.14.182",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.10",
@@ -53,8 +55,6 @@
   "dependencies": {
     "@floating-ui/react-dom": "^1.0.1",
     "@react-hook/debounce": "^4.0.0",
-    "@seed-ui/icons": "^0.6.0",
-    "@seed-ui/styles": "^0.10.0",
     "@vanilla-extract/dynamic": "^2.0.2",
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
