--- packages/elements/src/assets/icons.css
@@ -1,1 +1,0 @@
-@import "@seed-ui/icons/css/icons.css";
