--- packages/elements/src/assets/styles.css
@@ -1,1 +1,0 @@
-@import "@seed-ui/styles/css/styles.css";
