+++ packages/elements/src/components/Box/index.ts
@@ -1,0 +1,1 @@
+export * from './Box';
