+++ packages/elements/src/components/Button/Button.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex } from '@seed-ui/flexbox';
+import { Box, Flex } from '@seed-ui/elements';
 
 import Button, { ButtonProps, ButtonSize, ButtonVariant } from './Button';
