+++ packages/elements/src/components/DateInput/DateInput.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex } from '@seed-ui/flexbox';
+import { Box, Flex } from '@seed-ui/elements';
 
 import DateInput, { DateInputProps, DateInputSize } from './DateInput';
