+++ packages/elements/src/components/Flex/index.ts
@@ -1,0 +1,1 @@
+export * from './Flex';
