+++ packages/elements/src/components/IconButton/IconButton.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex } from '@seed-ui/flexbox';
+import { Box, Flex } from '@seed-ui/elements';
 
 import { Avatar } from '../Avatar';
