+++ packages/elements/src/components/Select/Select.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex } from '@seed-ui/flexbox';
+import { Box, Flex } from '@seed-ui/elements';
 
 import Select, { SelectProps, SelectSize } from './Select';
