+++ packages/elements/src/components/Textarea/Textarea.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex } from '@seed-ui/flexbox';
+import { Box, Flex } from '@seed-ui/elements';
 
 import Textarea, { TextareaProps, TextareaSize } from './Textarea';
