+++ packages/elements/src/index.ts
@@ -1,6 +1,3 @@
-export * from '@seed-ui/icons';
-export * from '@seed-ui/styles';
-
 // Styles
 export * from './styles';
 
@@ -9,10 +6,12 @@ export * from './components/Alert';
 export * from './components/Animation';
 export * from './components/Autocomplete';
 export * from './components/Avatar';
+export * from './components/Box';
 export * from './components/Button';
 export * from './components/Caption';
 export * from './components/DateInput';
 export * from './components/Divider';
+export * from './components/Flex';
 export * from './components/IconButton';
 export * from './components/InputGroup';
 export * from './components/Link';
@@ -32,6 +31,7 @@ export * from './components/Tag';
 export * from './components/Text';
 export * from './components/Textarea';
 export * from './components/TextInput';
+export * from './components/ThemeProvider';
 export * from './components/Title';
 export * from './components/Toast';
 export * from './components/ToastContainer';
