--- packages/flexbox/CHANGELOG.md
@@ -1,102 +1,0 @@
-# @seed-ui/flexbox
-
-## 0.2.6
-
-### Patch Changes
-
-- c9e6344: Improved color palette.
-- 6130817: Improved color palette.
-- Updated dependencies [c9e6344]
-- Updated dependencies [6130817]
-  - @seed-ui/styles@0.10.0
-
-## 0.2.5
-
-### Patch Changes
-
-- 23222ee: Added minified styled to packages.
-
-## 0.2.4
-
-### Patch Changes
-
-- 7ffe5d6: Added processed styles to packages.
-
-## 0.2.3
-
-### Patch Changes
-
-- ca02f98: Fixed rollup config.
-
-## 0.2.2
-
-### Patch Changes
-
-- b764ffd: Fixed rollup config.
-
-## 0.2.2
-
-### Patch Changes
-
-- 0365530: Included @vanila-extract/babel-plugin to the build.
-
-## 0.2.1
-
-### Patch Changes
-
-- 123d65a: Updated @vanilla-extract/css to 1.10.0
-
-## 0.2.0
-
-### Minor Changes
-
-- 2f18a2d: Migrated to React 18.
-
-### Patch Changes
-
-- Updated dependencies [2f18a2d]
-  - @seed-ui/styles@0.9.0
-
-## 0.1.5
-
-### Patch Changes
-
-- ad6dd9f:
-  - Made label tag default in the Label component.
-  - Updated Icon, IconButton, and Avatar components.
-
-## 0.1.4
-
-### Patch Changes
-
-- 89c40f5: Added dark variant to elements.
-
-## 0.1.3
-
-### Patch Changes
-
-- 82b8ec7: - Added default variant for link
-  - Fixed `lineHeight` property of Box component
-
-## 0.1.2
-
-### Patch Changes
-
-- Updated dependencies [32c592b]
-  - @seed-ui/styles@0.8.0
-
-## 0.1.1
-
-### Patch Changes
-
-- 545ff2d:
-  - Standardized elements sizes, so elements with same size look better in conjunction with each other
-  - Made all properties of flexbox components optional
-
-## 0.1.0
-
-### Minor Changes
-
-- e8260f9:
-  - Normalised buttons sizes
-  - Added flexbox package
