--- packages/flexbox/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@seed-ui/flexbox",
-  "version": "0.2.6",
-  "description": "",
-  "files": [
-    "css",
-    "dist",
-    "package.json",
-    "CHANGELOG.md"
-  ],
-  "types": "dist/types/index.d.ts",
-  "main": "dist/cjs/index.js",
-  "module": "dist/es/index.js",
-  "sideEffects": [
-    "*.css"
-  ],
-  "engines": {
-    "node": ">=16.x"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/smrwsky/seed-ui.git",
-    "directory": "packages/flexbox"
-  },
-  "scripts": {
-    "build": "yarn clean && rollup -c rollup.config.mjs",
-    "clean": "rimraf dist css",
-    "test": "jest --config jest.config.js --passWithNoTests",
-    "test:watch": "yarn test --watch",
-    "typecheck": "tsc -p tsconfig.json --noEmit"
-  },
-  "keywords": [],
-  "author": "Sergey Murawsky",
-  "license": "MIT",
-  "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.10",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "4.7.4"
-  },
-  "dependencies": {
-    "@seed-ui/styles": "^0.10.0",
-    "classnames": "^2.3.1",
-    "lodash": "^4.17.21"
-  }
-}
