--- packages/flexbox/src/components/Box/index.ts
@@ -1,2 +1,0 @@
-export * from './Box';
-export { default as Box } from './Box';
