--- packages/flexbox/src/components/Flex/index.ts
@@ -1,2 +1,0 @@
-export * from './Flex';
-export { default as Flex } from './Flex';
