--- packages/flexbox/src/index.ts
@@ -1,4 +1,0 @@
-export * from '@seed-ui/styles';
-
-export * from './components/Flex';
-export * from './components/Box';
