--- packages/icons/CHANGELOG.md
@@ -1,263 +1,0 @@
-# @seed-ui/icons
-
-## 0.6.7
-
-### Patch Changes
-
-- Updated dependencies [c9e6344]
-- Updated dependencies [6130817]
-  - @seed-ui/styles@0.10.0
-
-## 0.6.6
-
-### Patch Changes
-
-- 91a5de5:
-  - Fixed Toast styles
-  - Autocomplete improvements
-
-## 0.6.5
-
-### Patch Changes
-
-- 23222ee: Added minified styled to packages.
-
-## 0.6.4
-
-### Patch Changes
-
-- 7ffe5d6: Added processed styles to packages.
-
-## 0.6.3
-
-### Patch Changes
-
-- ca02f98: Fixed rollup config.
-
-## 0.6.2
-
-### Patch Changes
-
-- b764ffd: Fixed rollup config.
-
-## 0.6.2
-
-### Patch Changes
-
-- 0365530: Included @vanila-extract/babel-plugin to the build.
-
-## 0.6.1
-
-### Patch Changes
-
-- 123d65a: Updated @vanilla-extract/css to 1.10.0
-
-## 0.6.0
-
-### Minor Changes
-
-- 2f18a2d: Migrated to React 18.
-
-### Patch Changes
-
-- Updated dependencies [2f18a2d]
-  - @seed-ui/styles@0.9.0
-
-## 0.5.5
-
-### Patch Changes
-
-- ad6dd9f:
-  - Made label tag default in the Label component.
-  - Updated Icon, IconButton, and Avatar components.
-
-## 0.5.4
-
-### Patch Changes
-
-- 89c40f5: Added dark variant to elements.
-
-## 0.5.3
-
-### Patch Changes
-
-- Updated dependencies [32c592b]
-  - @seed-ui/styles@0.8.0
-
-## 0.5.2
-
-### Patch Changes
-
-- 545ff2d:
-  - Standardized elements sizes, so elements with same size look better in conjunction with each other
-  - Made all properties of flexbox components optional
-
-## 0.5.1
-
-### Patch Changes
-
-- e8260f9:
-  - Normalised buttons sizes
-  - Added flexbox package
-
-## 0.5.0
-
-### Minor Changes
-
-- e183122:
-  - Added new properties for theme
-  - Added new style utilities
-  - Standardised variant and props names
-  - Polished components' styles
-
-### Patch Changes
-
-- Updated dependencies [e183122]
-  - @seed-ui/styles@0.7.0
-
-## 0.4.8
-
-### Patch Changes
-
-- f00490f: Changed line height of icons.
-
-## 0.4.7
-
-### Patch Changes
-
-- cf68a61: Minor improvements.
-
-## 0.4.6
-
-### Patch Changes
-
-- 0bd5f3a: Minor improvements.
-
-## 0.4.5
-
-### Patch Changes
-
-- 2db0390: Fixed icon styles.
-
-## 0.4.4
-
-### Patch Changes
-
-- 6f20735: Fixed icon styles.
-
-## 0.4.3
-
-### Patch Changes
-
-- Updated dependencies [cdc01d6]
-  - @seed-ui/styles@0.6.0
-
-## 0.4.2
-
-### Patch Changes
-
-- e77506d: Added feedback components (Alert, Message, Toast);Improved Autocomplete and MenuLists;Added as property for InputGroup components;Added width properties for Button components;Updated bold styles
-
-## 0.4.1
-
-### Patch Changes
-
-- Updated dependencies [a8ce35f]
-  - @seed-ui/styles@0.5.0
-
-## 0.4.0
-
-### Minor Changes
-
-- fc8643d: Normalized icon's sizes.
-
-## 0.3.15
-
-### Patch Changes
-
-- Updated dependencies [6d5cc70]
-  - @seed-ui/styles@0.4.0
-
-## 0.3.14
-
-### Patch Changes
-
-- 9b56f05: Removed global css import from icon component.
-
-## 0.3.13
-
-### Patch Changes
-
-- da1c163: Replaced custom icons with boxicons.
-
-## 0.3.12
-
-### Patch Changes
-
-- ad72068: Created ImageIcon component.
-
-## 0.3.11
-
-### Patch Changes
-
-- cfd2f79: Fixed icons line-height.
-
-## 0.3.10
-
-### Patch Changes
-
-- eee8424: Updated package versions.
-
-## 0.3.9
-
-### Patch Changes
-
-- 0ff0601: Updated @seed-ui/styles package version for @seed-ui/layout and @seed-ui/icons to 0.3.5.
-
-## 0.3.8
-
-### Patch Changes
-
-- 8d9ea9e: Updated @seed-ui/styles package version in other packages.
-
-## 0.3.7
-
-### Patch Changes
-
-- c203b0d: Added border to design-tokens.
-
-## 0.3.6
-
-### Patch Changes
-
-- fbb4c21: Added babel plugin to transform svg icons to react components.
-
-## 0.3.5
-
-### Patch Changes
-
-- b95e4f9: Fixed broken build of @seed/ui/icons package.
-
-## 0.3.4
-
-### Patch Changes
-
-- b63dea7: Updated pachage versions.
-
-## 0.3.3
-
-### Patch Changes
-
-- 2ed6629: Added justifySelf property and classNames to Col component.
-
-## 0.3.2
-
-### Patch Changes
-
-- 06c3436: Added justifySelf atom styles.
-
-## 0.3.1
-
-### Patch Changes
-
-- 9ff2f24: Base components and styles created.
