--- packages/icons/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@seed-ui/icons",
-  "version": "0.6.7",
-  "description": "",
-  "files": [
-    "css",
-    "dist",
-    "fonts",
-    "package.json",
-    "CHANGELOG.md"
-  ],
-  "types": "dist/types/index.d.ts",
-  "main": "dist/cjs/index.js",
-  "module": "dist/es/index.js",
-  "sideEffects": [
-    "*.css"
-  ],
-  "engines": {
-    "node": ">=16.x"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/smrwsky/seed-ui.git",
-    "directory": "packages/icons"
-  },
-  "scripts": {
-    "build": "yarn clean && rollup -c rollup.config.mjs",
-    "clean": "rimraf dist css fonts",
-    "test": "jest --config jest.config.js --passWithNoTests",
-    "test:watch": "yarn test --watch",
-    "typecheck": "tsc -p tsconfig.json --noEmit"
-  },
-  "keywords": [],
-  "author": "Sergey Murawsky",
-  "license": "MIT",
-  "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@seed-ui/elements": "^0.10.0",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.10",
-    "@vanilla-extract/css": "^1.10.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "4.7.4"
-  },
-  "dependencies": {
-    "@seed-ui/styles": "^0.10.0",
-    "boxicons": "^2.1.2",
-    "classnames": "^2.3.1"
-  }
-}
