--- packages/icons/src/components/Icon/index.ts
@@ -1,2 +1,0 @@
-export * from './Icon';
-export { default as Icon } from './Icon';
