--- packages/icons/src/index.ts
@@ -1,4 +1,0 @@
-// Import external css at the top of the root file
-import 'boxicons/css/boxicons.css';
-
-export * from './components/Icon';
