--- packages/icons/src/types/global.d.ts
@@ -1,1 +1,0 @@
-export type Maybe<T> = T | null | undefined;
