+++ packages/styles/rollup.config.mjs
@@ -1,6 +1,7 @@
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import typescript from '@rollup/plugin-typescript';
 import { vanillaExtractPlugin } from '@vanilla-extract/rollup-plugin';
+import copy from 'rollup-plugin-copy';
 import del from 'rollup-plugin-delete'
 import externals from 'rollup-plugin-node-externals';
 import postcss from 'rollup-plugin-postcss';
@@ -77,11 +78,16 @@ export default [
     plugins: [
       nodeResolve(),
       externals({
-        exclude: [/minireset\.css/]
+        exclude: [/boxicons/, /minireset\.css/]
       }),
       postcss({
         extract: 'css/styles.css',
         sourceMap: true,
+      }),
+      copy({
+        targets: [
+          { src: '../../node_modules/boxicons/fonts/*', dest: 'fonts' },
+        ],
       }),
       del({
         targets: ['dist/es/**/*.css'],
@@ -104,6 +110,7 @@ export default [
         sourceMap: true,
       }),
     ],
+
     onwarn(warning, warn) {
       if (warning.code === 'FILE_NAME_CONFLICT') return
       warn(warning)
