+++ packages/styles/src/index.ts
@@ -1,12 +1,11 @@
 // Import external css at the top of the root file
 import 'minireset.css';
+import 'boxicons/css/boxicons.min.css';
 
 export * from './const';
 export * from './types';
 
-export * from './components/ThemeProvider';
-
-// Order of exported styles should be resolved
+// Order of exported styles should be preserved
 export * from './styles/mixins';
 export * from './styles/global.css';
 export * from './styles/atoms.css';
